Move config origin representers from /shared to merged_environments package
