+++ server/webapp/WEB-INF/rails.new/app/presenters/api_v1/admin/merged_environments/environment_variable_representer.rb
@@ -29,8 +29,8 @@ module ApiV1
                  exec_context: :decorator,
                  decorator: lambda {|origin, *|
                    (origin.instance_of? FileConfigOrigin) ?
-                     Shared::ConfigOrigin::ConfigXmlOriginRepresenter :
-                     Shared::ConfigOrigin::ConfigRepoOriginRepresenter
+                     ConfigOrigin::ConfigXmlOriginRepresenter :
+                     ConfigOrigin::ConfigRepoOriginRepresenter
                  }
 
         def origin
