+++ server/webapp/WEB-INF/rails.new/app/presenters/api_v1/admin/merged_environments/merged_environment_config_representer.rb
@@ -42,9 +42,9 @@ module ApiV1
                    },
                    decorator: lambda {|origin, *|
                      if origin.instance_of? FileConfigOrigin
-                       Shared::ConfigOrigin::ConfigXmlOriginRepresenter
+                       ConfigOrigin::ConfigXmlOriginRepresenter
                      else
-                       Shared::ConfigOrigin::ConfigRepoOriginRepresenter
+                       ConfigOrigin::ConfigRepoOriginRepresenter
                      end
                    }
