+++ server/webapp/WEB-INF/rails.new/spec/presenters/api_v1/admin/merged_environments/config_origin/config_repo_origin_representer_spec.rb
@@ -16,11 +16,11 @@
 
 require 'rails_helper'
 
-describe ApiV1::Shared::ConfigOrigin::ConfigRepoOriginRepresenter do
+describe ApiV1::Admin::MergedEnvironments::ConfigOrigin::ConfigRepoOriginRepresenter do
   it 'should render remote config origin with hal representation' do
     config_repo = ConfigRepoConfig.new(GitMaterialConfig.new('https://github.com/config-repos/repo', 'master'), 'json-plugin', 'repo1')
     config_repo_origin = RepoConfigOrigin.new(config_repo, 'revision1')
-    actual_json = ApiV1::Shared::ConfigOrigin::ConfigRepoOriginRepresenter.new(config_repo_origin).to_hash(url_builder: UrlBuilder.new)
+    actual_json = ApiV1::Admin::MergedEnvironments::ConfigOrigin::ConfigRepoOriginRepresenter.new(config_repo_origin).to_hash(url_builder: UrlBuilder.new)
 
     expect(actual_json).to eq(expected_json)
   end
