+++ server/webapp/WEB-INF/rails.new/spec/presenters/api_v1/admin/merged_environments/config_origin/config_xml_origin_representer_spec.rb
@@ -16,9 +16,9 @@
 
 require 'rails_helper'
 
-describe ApiV1::Shared::ConfigOrigin::ConfigXmlOriginRepresenter do
+describe ApiV1::Admin::MergedEnvironments::ConfigOrigin::ConfigXmlOriginRepresenter do
   it 'should render local config origin' do
-    presenter = ApiV1::Shared::ConfigOrigin::ConfigXmlOriginRepresenter.new(get_config_xml_origin)
+    presenter = ApiV1::Admin::MergedEnvironments::ConfigOrigin::ConfigXmlOriginRepresenter.new(get_config_xml_origin)
     actual_json = presenter.to_hash(url_builder: UrlBuilder.new)
     expect(actual_json).to eq(expected_json)
   end
